+++ jest.config.js
@@ -90,11 +90,14 @@ module.exports = {
   // A map from regular expressions to module names or to arrays of module names that allow to stub out resources with a single module
   // moduleNameMapper: {},
   moduleNameMapper: {
+    '^@subql/common-substrate$': '<rootDir>/packages/common-substrate/src/',
     '^@subql/common-substrate/(.*)$': '<rootDir>/packages/common-substrate/src/$1',
+    '^@subql/common$': '<rootDir>/packages/common/src/',
     '^@subql/common/(.*)$': '<rootDir>/packages/common/src/$1',
+    '^@subql/node-core$': '<rootDir>/packages/node-core/src/',
     '^@subql/node-core/(.*)$': '<rootDir>/packages/node-core/src/$1',
+    '^@subql/utils$': '<rootDir>/packages/utils/src/',
     '^@subql/utils/(.*)$': '<rootDir>/packages/utils/src/$1',
-    '^@subql/node-core/logger': '<rootDir>/packages/node-core/logger',
   },
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader