+++ packages/node-core/src/indexer/dictionary.service.test.ts
@@ -112,7 +112,6 @@ const nodeConfig = new NodeConfig({
 describe('DictionaryService', () => {
   it('return dictionary query result', async () => {
     const dictionaryService = new DictionaryService(DICTIONARY_ENDPOINT, DICTIONARY_CHAINID, nodeConfig);
-    await dictionaryService.init();
 
     const batchSize = 30;
     const startBlock = 1;
@@ -128,7 +127,6 @@ describe('DictionaryService', () => {
       '0x21121',
       nodeConfig
     );
-    await dictionaryService.init();
 
     const batchSize = 30;
     const startBlock = 1;
@@ -139,7 +137,6 @@ describe('DictionaryService', () => {
 
   it('should return meta even startblock height greater than dictionary last processed height', async () => {
     const dictionaryService = new DictionaryService(DICTIONARY_ENDPOINT, DICTIONARY_CHAINID, nodeConfig);
-    await dictionaryService.init();
 
     const batchSize = 30;
     const startBlock = 400000000;
@@ -150,7 +147,6 @@ describe('DictionaryService', () => {
 
   it('test query the correct range', async () => {
     const dictionaryService = new DictionaryService(DICTIONARY_ENDPOINT, DICTIONARY_CHAINID, nodeConfig);
-    await dictionaryService.init();
 
     const batchSize = 30;
     const startBlock = 1;
@@ -169,7 +165,6 @@ describe('DictionaryService', () => {
 
   it('use minimum value of event/extrinsic returned block as batch end block', async () => {
     const dictionaryService = new DictionaryService(DICTIONARY_ENDPOINT, DICTIONARY_CHAINID, nodeConfig);
-    await dictionaryService.init();
 
     const batchSize = 50;
     const startBlock = 333300;
@@ -216,9 +211,8 @@ describe('DictionaryService', () => {
     expect(dic?.batchBlocks[dic.batchBlocks.length - 1]).toBe(339186);
   }, 500000);
 
-  it('able to build queryEntryMap', async () => {
+  it('able to build queryEntryMap', () => {
     const dictionaryService = new DictionaryService(DICTIONARY_ENDPOINT, DICTIONARY_CHAINID, nodeConfig);
-    await dictionaryService.init();
 
     dictionaryService.buildDictionaryEntryMap(mockDS, () => HAPPY_PATH_CONDITIONS);
     const _map = (dictionaryService as any).mappedDictionaryQueryEntries;
@@ -227,9 +221,8 @@ describe('DictionaryService', () => {
     expect(_map.size).toEqual(mockDS.length);
   });
 
-  it('able to getDicitonaryQueryEntries', async () => {
+  it('able to getDicitonaryQueryEntries', () => {
     const dictionaryService = new DictionaryService(DICTIONARY_ENDPOINT, DICTIONARY_CHAINID, nodeConfig);
-    await dictionaryService.init();
     const dictionaryQueryMap = new Map();
 
     // Mocks a Map object that where key == dataSource.startBlock and mocked DictionaryQueryEntries[] values
@@ -267,9 +260,8 @@ describe('DictionaryService', () => {
     expect(dictionaryService.getDictionaryQueryEntries(queryEndBlock)).toEqual([]);
   });
 
-  it('sort map', async () => {
+  it('sort map', () => {
     const dictionaryService = new DictionaryService(DICTIONARY_ENDPOINT, DICTIONARY_CHAINID, nodeConfig);
-    await dictionaryService.init();
 
     const unorderedDs = [mockDS[2], mockDS[0], mockDS[1]];
     dictionaryService.buildDictionaryEntryMap(unorderedDs, (startBlock) => startBlock as any);