+++ packages/node/src/indexer/fetch.module.ts
@@ -102,15 +102,7 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
     },
     FetchService,
     BenchmarkService,
-    {
-      provide: DictionaryService,
-      useFactory: async (project: SubqueryProject, nodeConfig: NodeConfig) => {
-        const dictionaryService = new DictionaryService(project, nodeConfig);
-        await dictionaryService.init();
-        return dictionaryService;
-      },
-      inject: ['ISubqueryProject', NodeConfig],
-    },
+    DictionaryService,
     SandboxService,
     DsProcessorService,
     DynamicDsService,