+++ packages/node/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Unreleased -2023-06-16
+### Added
+- Support for transaction filter `to: null` (#108)
+
 ## [2.6.0] - 2023-06-15
 ### Added
 - Implement multiple-endpoint improvements from node-core (#102)